% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_orgaos}
\alias{get_orgaos}
\title{Extracts party organ datas from TSE}
\usage{
get_orgaos(agreg, partido = 0, zona = 0, uni_eleitoral = NULL,
  uni_eleitoral_superior = NULL, history = "false")
}
\arguments{
\item{agreg}{character - regional agregation desired. (nacional, estadual, municipal, regional(DF), zona(DF)}

\item{partido}{integer - party numver. 0 returns data from all parties.}

\item{zona}{integer - zone number. Only used with zona(DF) regional agregation.}

\item{uni_eleitoral}{character or integer - Electoral unity. Should not be used with nacional agregation. 
When used with "estadual" regional agregation, a state should be specified (ex. "SP", "RJ", "BA", etc.).
When used with "municipal" regional agregation, the municipality TSE number should be specified.}

\item{uni_eleitoral_superior}{character - Only used with "municipal" regional agregation. The state should be specified (ex. "SP", "RJ", "BA", etc.).}

\item{history}{character - false or true. If true, the API request will return data from previous years.}
}
\description{
A função get_orgaos extrai os orgãos partidádios de acordo com o nível regional utilizado pelo usuário.

* \code{get_orgaos()} extracts information about the party organ.

* \code{get_members()} extracts members informations from a specified organ.
}
